Init the notebook for reading in the data from TSV to the AnnData format.
This is largely transcribing and tweaking Dean Lee's code from the F1L
repo, just transliterated into marimo's notebook. To mitigate high memory
usage, I wrote a helper function to read and assemble the sparse matrix
in chunks.

Testing now, seems to run through with a subset of the data, but haven't
done the conversion to AnnData with the full dataset